Moved `PublicProvider` interface to core types module
